requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
google-ai-generativelanguage>=0.4.0
google-generativeai>=0.3.2
python-docx>=0.8.11